fix type

+++ api/售后API-售后列表.go
@@ -38,19 +38,19 @@ type (
 		PageNo              int64 `json:"pageNo"`
 		PageSize            int64 `json:"pageSize"`
 		AfterSaleBasicInfos []struct {
-			ReturnsId                string `json:"returnsId"`
-			ReturnType               int64  `json:"returnType"`
-			ReasonId                 int64  `json:"reasonId"`
-			ReasonNameZh             string `json:"reasonNameZh"`
-			Status                   int64  `json:"status"`
-			UserId                   string `json:"userId"`
-			OrderId                  string `json:"orderId"`
-			ApplyTime                int64  `json:"applyTime"`
-			UpdatedAt                int64  `json:"updatedAt"`
-			ExpireTime               int64  `json:"expireTime"`
-			Desc                     string `json:"desc"`
-			ReturnsTag               int64  `json:"returnsTag"`
-			ExpectedRefundAmountYuan int64  `json:"expectedRefundAmountYuan"`
+			ReturnsId                string  `json:"returnsId"`
+			ReturnType               int64   `json:"returnType"`
+			ReasonId                 int64   `json:"reasonId"`
+			ReasonNameZh             string  `json:"reasonNameZh"`
+			Status                   int64   `json:"status"`
+			UserId                   string  `json:"userId"`
+			OrderId                  string  `json:"orderId"`
+			ApplyTime                int64   `json:"applyTime"`
+			UpdatedAt                int64   `json:"updatedAt"`
+			ExpireTime               int64   `json:"expireTime"`
+			Desc                     string  `json:"desc"`
+			ReturnsTag               int64   `json:"returnsTag"`
+			ExpectedRefundAmountYuan float64 `json:"expectedRefundAmountYuan"`
 		} `json:"afterSaleBasicInfos"`
 		MaxPageNo int64 `json:"maxPageNo"`
 	}
